+++ plugin-document/src/test/java/nl/xillio/xill/plugins/mongodb/services/serializers/MongoBinarySerializerTest.java
@@ -15,6 +15,7 @@
  */
 package nl.xillio.xill.plugins.mongodb.services.serializers;
 
+import com.google.inject.Guice;
 import nl.xillio.xill.TestUtils;
 import nl.xillio.xill.api.components.MetaExpression;
 import nl.xillio.xill.api.io.SimpleIOStream;
@@ -29,11 +30,7 @@ import java.util.Base64;
 import static org.testng.Assert.*;
 
 public class MongoBinarySerializerTest extends TestUtils {
-    private MongoSerializer serializer = new MongoSerializer(
-            new ObjectIdSerializer(),
-            new UUIDSerializer(),
-            new MongoRegexSerializer(),
-            new BinarySerializer());
+    private MongoSerializer serializer = Guice.createInjector().getInstance(MongoSerializer.class);
 
     @Test
     public void testParseStringStream() throws IOException {